(PC-29870) refactor(favorites): Extract animation hook to avoid duplicate code
